Move test-fixtures and test-utils under __tests__
