+++ src/atoms/__tests__/paneActions.test.ts
@@ -6,12 +6,12 @@ import { describe, expect, it } from 'vitest';
 import { readFileContent, writeFileContent } from '../../io/filesystem';
 import { activePaneAtom, openFileEntryAtom } from '../editorActions';
 import { activePaneContentAtom, focusPaneAtom } from '../paneActions';
-import { makeFile } from '../test-fixtures';
-import { runGetAtomHook, runSetAtomHook } from '../test-utils';
 import { EditorContent } from '../types/EditorContent';
 import { EditorContentAtoms } from '../types/EditorContentAtoms';
 import { FileEntry } from '../types/FileEntry';
 import { PaneId } from '../types/PaneGroup';
+import { makeFile } from './test-fixtures';
+import { runGetAtomHook, runSetAtomHook } from './test-utils';
 
 vi.mock('../../io/filesystem');
