+++ src/components/__tests__/EditorsList.test.tsx
@@ -1,6 +1,6 @@
 import { VscFile } from 'react-icons/vsc';
 
-import { makeFile } from '../../atoms/test-fixtures';
+import { makeFile } from '../../atoms/__tests__/test-fixtures';
 import { noop } from '../../lib/noop';
 import { render, screen, setup } from '../../test/test-utils';
 import { EditorsList } from '../EditorsList';
