+++ src/features/references/__tests__/test-fixtures.ts
@@ -1,4 +1,4 @@
-import { ReferenceItem } from '../../types/ReferenceItem';
+import { ReferenceItem } from '../../../types/ReferenceItem';
 
 export const REFERENCES: ReferenceItem[] = [
   {
