+++ src/features/references/editor/__tests__/ReferencesTableView.test.tsx
@@ -1,8 +1,8 @@
+import { runSetAtomHook } from '../../../../atoms/__tests__/test-utils';
 import { setReferencesAtom } from '../../../../atoms/referencesState';
-import { runSetAtomHook } from '../../../../atoms/test-utils';
 import { act, screen, setupWithJotaiProvider } from '../../../../test/test-utils';
+import { REFERENCES } from '../../__tests__/test-fixtures';
 import { UploadTipInstructions } from '../../components/UploadTipInstructions';
-import { REFERENCES } from '../../test-fixtures';
 import { ReferencesTableView } from '../ReferencesTableView';
 
 describe('ReferencesTableView component', () => {
