+++ src/features/textEditor/components/tipTapNodes/__tests__/test-utils.test.ts
@@ -1,7 +1,7 @@
 import { Editor } from '@tiptap/react';
 
 import { EDITOR_EXTENSIONS } from '../../tipTapEditorConfigs';
-import { getPrettyHTML, getPrettyHTMLWithSelection, getSelectedText, setUpEditorWithSelection } from '../test-utils';
+import { getPrettyHTML, getPrettyHTMLWithSelection, getSelectedText, setUpEditorWithSelection } from './test-utils';
 
 describe('TipTap test utils', () => {
   const editor = new Editor({
