+++ src/features/textEditor/components/tipTapNodes/collapsibleBlock/__tests__/CollapsibleBlock.test.tsx
@@ -2,8 +2,8 @@ import { Editor, EditorContent } from '@tiptap/react';
 
 import { act, screen, setup } from '../../../../../../test/test-utils';
 import { EDITOR_EXTENSIONS } from '../../../tipTapEditorConfigs';
-import { defaultCollapsibleBlock, defaultUncollapsedCollapsibleBlock } from '../../test-fixtures';
-import { findNodesByNodeType } from '../../test-utils';
+import { defaultCollapsibleBlock, defaultUncollapsedCollapsibleBlock } from '../../__tests__/test-fixtures';
+import { findNodesByNodeType } from '../../__tests__/test-utils';
 
 describe('CollapsibleBlock', () => {
   const editor = new Editor({
