+++ src/features/textEditor/components/tipTapNodes/collapsibleBlock/helpers/__tests__/changeCollapsibleBlockToParagraphs.test.ts
@@ -2,8 +2,8 @@ import { TextSelection } from '@tiptap/pm/state';
 import { Editor, getText } from '@tiptap/react';
 
 import { EDITOR_EXTENSIONS } from '../../../../tipTapEditorConfigs';
-import { defaultCollapsibleBlock } from '../../../test-fixtures';
-import { findNodesByNodeType } from '../../../test-utils';
+import { defaultCollapsibleBlock } from '../../../__tests__/test-fixtures';
+import { findNodesByNodeType } from '../../../__tests__/test-utils';
 import { changeCollapsibleBlockToParagraphs } from '../changeCollapsibleBlockToParagraphs';
 
 describe('changeCollapsibleBlockToParagraph', () => {
