+++ src/features/textEditor/components/tipTapNodes/collapsibleBlock/helpers/__tests__/unsetPartiallySelectedCollapsibleBlocks.test.ts
@@ -1,8 +1,8 @@
 import { Editor } from '@tiptap/react';
 
 import { EDITOR_EXTENSIONS } from '../../../../tipTapEditorConfigs';
-import { defaultCollapsibleBlock } from '../../../test-fixtures';
-import { getPrettyHTMLWithSelection, setUpEditorWithSelection } from '../../../test-utils';
+import { defaultCollapsibleBlock } from '../../../__tests__/test-fixtures';
+import { getPrettyHTMLWithSelection, setUpEditorWithSelection } from '../../../__tests__/test-utils';
 import { unsetPartiallySelectedCollapsibleBlocks } from '../unsetPartiallySelectedCollapsibleBlocks';
 
 describe('unsetPartiallySelectedCollapsibleBlocks', () => {
